Fix concurrent make
The Makefile contained two race conditions that could mess up concurrent
builds with many processes.

1) Fontmake is hardcoded to put binaries into a master_* directory, the
Makefile tried to move them out of there and then delete the directory.
This failed when concurrent makes were writing to the same directory.

2) The Makefile tried to move the binary by taking the last modified
file out of the directory. This led to bogus results when multiple
processes were writing to the directory concurrently.

Fix it by not removing the master_ttf directory and using a script to
find the name of the binary that fontmake would build.

Git will not pick up the master_ttf directory, so no modification of
.gitignore is needed.