Set version to 1.000, provide version setter script, update composites
The baseline release is supposed to match the Google Fonts release as
closely as possible. Since the GF release has been used in the wild for
years, it is appropriate to set the version to 1.000 and then go from
there.

The version script changes the metadata as per the recommendations in
http://silnrsi.github.io/FDBP/en-US/Versioning.html and additionally
updates the version glyph (uniEFFD).

Since the version glyph was previously inconsistent and contained
components and sometimes decomposed ones, I had to update the composite
information with `python -m vttLib dump font.ttf font.ufo` before VTT
compilation succeeded. Due to spacing changes, there is a large amount
of churn.